Mapping: Refactoring geocoder and loader
