--- dt-mapping/geocode-api/geocoder.php
@@ -1,1 +1,0 @@
-<?php
